+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+## [1.4.0](https://github.com/tmchow/energon/compare/v1.3.0...v1.4.0) (2026-09-15)
+
+
+### Features
+
+* **skill:** land routine upstream updates by repository policy ([#117](https://github.com/tmchow/energon/issues/117)) ([4933ee5](https://github.com/tmchow/energon/commit/4933ee551a2032b5f3770aa190fd832cbf718ddd))
+
+## [1.3.0](https://github.com/tmchow/energon/compare/v1.2.0...v1.3.0) (2026-09-15)
+
+
+### Features
+
+* **skill:** add deploy-this-energon operator skill ([#115](https://github.com/tmchow/energon/issues/115)) ([ecd86c1](https://github.com/tmchow/energon/commit/ecd86c12533e87cc4a840501b3056eba186a8139))
+
+## [1.2.0](https://github.com/tmchow/energon/compare/v1.1.0...v1.2.0) (2026-09-15)
+
+
+### Features
+
+* **skill:** add fork update and backup operator skills ([#112](https://github.com/tmchow/energon/issues/112)) ([d094239](https://github.com/tmchow/energon/commit/d094239a5181fe93b3e2604e21b95cf9421e2c5d))
+
+
+### Bug Fixes
+
+* **deploy:** support independent private deployment repositories ([#113](https://github.com/tmchow/energon/issues/113)) ([0d200f3](https://github.com/tmchow/energon/commit/0d200f31249fce3e896fa8e8132f6717d94e486b))
+
+## [1.1.0](https://github.com/tmchow/energon/compare/v1.0.0...v1.1.0) (2026-09-15)
+
+
+### Features
+
+* **hub:** show when an upstream release is available ([#108](https://github.com/tmchow/energon/issues/108)) ([e42c5c6](https://github.com/tmchow/energon/commit/e42c5c6d690efdab86a947130aaf8a5cec9625c1))
+
 ## 1.0.0 (2026-09-14)
 
 Initial public release.